revised jumping code to be more responsive

+++ core/src/com/brassgames/entities/Entity.java
@@ -21,7 +21,7 @@ public abstract class Entity {
 	}
 
 	
-	public abstract void update(long delta, KeyboardListener keyboard);
+	public abstract void update(float delta, KeyboardListener keyboard);
 	
 	public abstract void render(SpriteBatch batch);
 }
